Combine create and update states into present
This combines the create and update states into a single present state.
The main reason for doing this is so the playbook author does not need
to know anything about the existing state of the resource. If the
resource does not already exist, we will create it. If it does exist, we
will update it to reflect the provided desired state.