Move collection files to package data dir
The module_utils and unit tests covering module_utils are not actually
part of the project's code. They are just files that get copied to the
collection. It makes more sense to keep these separate from the main
code. This adds a "data" directory to the package that mirrors the
structure of the amazon.cloud collection. This directory is copied in
its entirety to the generated collection.