Combine create and update states into present (#8)
Combine create and update states into present

Depends-On: #9
SUMMARY

This combines the create and update states into a single present state.
The main reason for doing this is so the playbook author does not need
to know anything about the existing state of the resource. If the
resource does not already exist, we will create it. If it does exist, we
will update it to reflect the provided desired state.

ISSUE TYPE


Feature Pull Request

COMPONENT NAME

ADDITIONAL INFORMATION

Reviewed-by: Alina Buzachis <None>