dedicated command to collect the schema definitions (#55)
dedicated command to collect the schema definitions

refresh_schema is in charge to retrieve the schema from AWS.
refresh_modules now reads a local copy of the schema definitions.

This allow us to:

run refresh_modules in the CI without any AWS credentials.
harmonize the way the module generators work

Reviewed-by: Alina Buzachis <None>
Reviewed-by: Gonéri Le Bouder <goneri@lebouder.net>